requirements-dev.txt
Django>=3.0
PyJWT>=2.3
cryptography>=36.0
requests>=2.27